+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.19
 
+* New runtime option for JES: `bootDiskSizeGb`. Allows specification of a boot disk size (as an Integer number of GB) that can be increased to boot a larger docker image.
 * Bug fix: Tasks that changed directory would fail on JES because their return code file was written to the new directory instead of an absolute path
 * Bug fix: Using `write_*` functions in a Task's command (e.g. `./my_script --file=${write_file(my_array)}`) will now work with JES
 * Changing format of the 'disks' runtime attribute slightly to allow for mounting disks at specific mountpoints